Switch from a 32 bit to 16 bit floating point buffer to improve performance.

+++ OpenGLStudy/core/PingPongBuffer.cpp
@@ -25,7 +25,7 @@ void PingPongBuffer::addTextureChannel(unsigned int textureUnit, const float * b
 	for (int i = 0; i < 2; i++)
 	{
 		glBindTexture(GL_TEXTURE_2D, newChannel->texture[i]);
-		glTexImage2D(GL_TEXTURE_2D, 0, GL_RGBA32F, width, height, 0, GL_RGBA, GL_FLOAT, NULL);
+		glTexImage2D(GL_TEXTURE_2D, 0, GL_RGBA16F, width, height, 0, GL_RGBA, GL_FLOAT, NULL);
 		glTexParameteri(GL_TEXTURE_2D, GL_TEXTURE_WRAP_S, GL_CLAMP_TO_BORDER);
 		glTexParameteri(GL_TEXTURE_2D, GL_TEXTURE_WRAP_T, GL_CLAMP_TO_BORDER);
 		glTexParameterfv(GL_TEXTURE_2D, GL_TEXTURE_BORDER_COLOR, borderValues);
